# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API ePortfolio",
      "description": "Daftar API ePortfolio",
      "version": "1.0"
    },
    "servers":[
        {
            "url": "http://127.0.0.1:8800/api",
            "description":"localhost"
        },
        {
            "url": "http://103.187.223.15:8800/api",
            "description":"dev"
        }
    ],
...
    "produces": ["application/json"],
    "paths": {
...
      "/articles/":{
        "post":{
          "tags":[
            "Article"
          ],
          "summary":"Create Article",
          "description":"Create a new article",
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    },
                    "coverArticle":{
                      "type":"string"
                    },
                    "title":{
                      "type":"string"
                    },
                    "desc":{
                      "type":"string"
                    },
                    "mermaidDiagram":{
                      "type":"string"
                    },
                    "isPublic":{
                      "type":"boolean"
                    },
                    "tags":{
                      "type":"array",
                      "items":{
                        "type":"string"
                      }
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/articles/{id}":{
        "put":{
          "tags":[
            "Article"
          ],
          "summary":"Update Article",
          "description":"Update article",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"article id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    },
                    "coverArticle":{
                      "type":"string"
                    },
                    "title":{
                      "type":"string"
                    },
                    "desc":{
                      "type":"string"
                    },
                    "mermaidDiagram":{
                      "type":"string"
                    },
                    "isPublic":{
                      "type":"boolean"
                    },
                    "tags":{
                      "type":"array",
                      "items":{
                        "type":"string"
                      }
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "403":{
              "description":"Forbidden"
            },
            "500":{
              "description": "Error"
            }
          }
        },
        "delete":{
          "tags":[
            "Article"
          ],
          "summary":"Delete Article",
          "description":"Delete article",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"article id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "403":{
              "description":"Forbidden"
            },
            "500":{
              "description": "Error"
            }
          }
        },
        "get":{
          "tags":[
            "Article"
          ],
          "summary":"Get Article",
          "description":"Get article data",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"article id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"Article not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/articles/all/{id}":{
        "get":{
          "tags":[
            "Article"
          ],
          "summary":"Get All Articles From User",
          "description":"Get all articles data from user",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"user id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/articles/timeline/all/{id}":{
        "get":{
          "tags":[
            "Article"
          ],
          "summary":"Get All Articles for home",
          "description":"Get all articles data for home, only articles from user who followed",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"user id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/articles/search/{keyword}":{
        "get":{
          "tags":[
            "Article"
          ],
          "summary":"Get Post by Keyword",
          "description":"Get post data by keyword",
          "parameters":[
            {
              "in":"path",
              "name": "keyword",
              "description":"keyword"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
...
      "/badges":{
        "post":{
          "tags":[
            "Badge"
          ],
          "summary":"Create badge",
          "description":"Create a new badge",
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    },
                    "title":{
                      "type":"string"
                    },
                    "imgBadge":{
                      "type":"string"
                    },
                    "issuer":{
                      "type":"string"
                    },
                    "desc":{
                      "type":"string"
                    },
                    "url":{
                      "type":"string"
                    },
                    "earnedDate":{
                      "type":"string"
                    },
                    "skills":{
                      "type":"array",
                      "items":{
                        "type":"string"
                      }
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/badges/{id}":{
        "put":{
          "tags":[
            "Badge"
          ],
          "summary":"Update Badge",
          "description":"Update badge",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"badge id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    },
                    "title":{
                      "type":"string"
                    },
                    "imgBadge":{
                      "type":"string"
                    },
                    "issuer":{
                      "type":"string"
                    },
                    "desc":{
                      "type":"string"
                    },
                    "url":{
                      "type":"string"
                    },
                    "earnedDate":{
                      "type":"string"
                    },
                    "skills":{
                      "type":"array",
                      "items":{
                        "type":"string"
                      }
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "403":{
              "description":"Forbidden"
            },
            "404":{
              "description":"Project not found"
            },
            "500":{
              "description": "Error"
            }
          }
        },
        "delete":{
          "tags":[
            "Badge"
          ],
          "summary":"Delete Badge",
          "description":"Delete badge",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"badge id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "403":{
              "description":"Forbidden"
            },
            "404":{
              "description":"Project not found"
            },
            "500":{
              "description": "Error"
            }
          }
        },
        "get":{
          "tags":[
            "Badge"
          ],
          "summary":"Get Badge",
          "description":"Get badge data",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"badge id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"Badge not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/badges/all/{id}":{
        "get":{
          "tags":[
            "Badge"
          ],
          "summary":"Get All Badges From User",
          "description":"Get all badges data from user",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"user id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
...
    }
 }